# backend/src/main/java/com/example/hospitalscheduler/dto/LoginRequest.java
package com.example.hospitalscheduler.dto;

public class LoginRequest {
    private String rut;
    private String password;

    // Getters and Setters
    public String getRut() {
        return rut;
    }

    public void setRut(String rut) {
        this.rut = rut;
    }

    public String getPassword() {
        return password;
    }

    public void setPassword(String password) {
        this.password = password;
    }
} 